+++ core/setup_scripts/check_configuration.d/07-check_terminal_emulators.sh
@@ -63,7 +63,6 @@ for CURRENT_COMMAND in $TERMINAL_EMULATORS; do
       COMMAND_DIR=${COMMAND_DIR##* }
       if [ ! -z "$COMMAND_DIR" -a -f "$COMMAND_DIR" ]; then
          echo -n "found, but not in PATH"
-         ADD_TO_PATH=1
       fi
    else
       echo -n "found"
@@ -77,14 +76,6 @@ for CURRENT_COMMAND in $TERMINAL_EMULATORS; do
          # located. Moreover, such directory has not been inserted into the
          # list of candidate user signalled paths (PATH_TO_BE_USED).
          PATH_TO_BE_USED="$PATH_TO_BE_USED$COMMAND_PATH:"
-      fi
-      if [ "$FIXMODE" = "1" -a $ADD_TO_PATH -eq 1 ]; then
-         # Clean the PATH_TO_BE_USED variable, in order to suppress
-         # warnings concerning unavailability of the current tool.
-         PATH_TO_BE_USED=":"
-         ln -fs "$COMMAND_DIR" "$NETKIT_HOME/bin"
-         echo -n " (fixed)"
-         OK=1
       fi
       echo
    else